requirements.txt
numpy>=1.24.0
PyMuPDF>=1.23.0
google-generativeai>=0.3.0
python-docx>=1.1.0
Pillow>=10.0.0
python-dotenv>=1.0.0
docx2pdf>=0.1.8
fpdf2>=2.5.5
pytesseract>=0.3.10
pdf2image>=1.16.3
pyinstaller>=5.13.0
pydicom>=2.6.1
pymongo>=4.6.0
certifi>=2024.0.0
rpds-py>=0.14.0
chromadb>=0.4.22
PyJWT>=2.8.0
cryptography>=41.0.0
requests>=2.31.0
pandas>=2.0.0
plotly>=5.17.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6